scipubs_mas_cli/requirements.txt
psycopg2-binary
pydantic>=2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
requests
matplotlib
